refactor: Fixtures
